Add files via upload
The pdf file is the export price from DHL(th)
The objective for retrieving and generating this pdf file to Dataframe is for my own interest

You might see that the cleaning dataframe and the read pdf is not related due to I have done it in different file.
I only merge code into one single file. If you have any question related to it, feel free to ask.